+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* added cost info and query stats
+* added query stats
 
 ## 1.10.0
 * changed level of update token log record from info to debug